Make production mesh behavior auditable and canonical
Flatten single-child modules, centralize strict parsing and validation, preserve canonical Method-C precision, and make GUI/runtime capabilities single-source.

Constraint: Preserve canonical mixed-precision physics and existing production data contracts
Rejected: Per-callsite guards and compatibility facades | they leave sibling paths inconsistent and retain architecture debt
Confidence: high
Scope-risk: broad
Directive: Keep namelist rewrites quote-aware and Method-C default-real boundaries explicit
Tested: make release-check; make test-gui; make clippy-full clippy-gui; make test-real-hydro; targeted namelist, restart, Gridfile, and 1/2/4-thread determinism tests
Not-tested: Complete make test-full replay after the final test-only thread-count assertion

+++ rust/earthmesh_cli/src/mesh_conversion_gridfile_state/from_state.rs
@@ -16,14 +16,14 @@ pub fn gridfile_mesh_from_state(grid: &GridMemory, tabs: &IjTabs) -> io::Result<
 
     let m_points = (0..nma)
         .map(|idx| LonLatPoint {
-            lon: f64::from(grid.glonm[idx]),
-            lat: f64::from(grid.glatm[idx]),
+            lon: grid.glonm[idx],
+            lat: grid.glatm[idx],
         })
         .collect();
     let w_points = (0..nwa)
         .map(|idx| LonLatPoint {
-            lon: f64::from(grid.glonw[idx]),
-            lat: f64::from(grid.glatw[idx]),
+            lon: grid.glonw[idx],
+            lat: grid.glatw[idx],
         })
         .collect();
     let m_to_w = tabs.m.iter().take(nma).map(|tab| tab.iw).collect();
@@ -72,14 +72,14 @@ pub fn gridfile_mesh_from_one_based_state(
 
     let m_points = (1..=nma)
         .map(|idx| LonLatPoint {
-            lon: f64::from(grid.glonm[idx]),
-            lat: f64::from(grid.glatm[idx]),
+            lon: grid.glonm[idx],
+            lat: grid.glatm[idx],
         })
         .collect();
     let w_points = (1..=nwa)
         .map(|idx| LonLatPoint {
-            lon: f64::from(grid.glonw[idx]),
-            lat: f64::from(grid.glatw[idx]),
+            lon: grid.glonw[idx],
+            lat: grid.glatw[idx],
         })
         .collect();
     let m_to_w = (1..=nma).map(|idx| tabs.m[idx].iw).collect();
